+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Adds support to one-call buy an order
 * Lowered the default timeout of requests from 120 seconds to 60 seconds
 * Added the Nodejs version in use to the User-Agent header on requests
+* Added support for create list of trackers
 
 ## 4.0.0 2021-10-06